refactor: Merging LLM Model & CausalLM  (#245)
